+++ src/components/Common/TopLoading/TopLoading.stories.tsx
@@ -1,21 +1,17 @@
-import { ComponentStory, ComponentMeta } from "@storybook/react";
-import React from "react";
+import { ComponentMeta, ComponentStory } from "@storybook/react";
 
 import { TopLoading } from "./TopLoading";
 
 export default {
-  title: "Common//TopLoading",
-  component: TopLoading,
-  tags: ['autodocs'],
-  parameters: {
-    layout: 'fullscreen',
-  },
-  argTypes: {
-  },
+    title: "Common/TopLoading",
+    component: TopLoading,
+    tags: ["autodocs"],
+    parameters: {
+        layout: "fullscreen",
+    },
+    argTypes: {},
 } as ComponentMeta<typeof TopLoading>;
 
-const Template: ComponentStory<typeof TopLoading> = () => <TopLoading  />;
+const Template: ComponentStory<typeof TopLoading> = () => <TopLoading />;
 export const Default = Template.bind({});
-Default.args = {
-
-};
+Default.args = {};
